Move MemStats struct to searchlib test library. Add class comment. Stop wrapping EntryType in WrappedEntry. Change Compare to reference value.  Add class comment. Remove WrappedCompare. Add freeze() method.
