+++ searchlib/src/vespa/searchlib/datastore/unique_store.h
@@ -24,32 +24,24 @@ public:
     using DataStoreType = DataStoreT<RefT>;
     using EntryType = EntryT;
     using RefType = RefT;
-    class WrappedEntry {
-        EntryType _value;
-    public:
-        WrappedEntry() : _value() { }
-        WrappedEntry(const EntryType &value_) : _value(value_) { }
-        WrappedEntry(const WrappedEntry &rhs) : _value(rhs.value()) { }
-        const EntryType &value() const { return _value; }
-    };
+    /*
+     * Compare two values in data store based on reference.  Invalid
+     * reference is mapped to local value reference to support
+     * comparing with new value candidate outside data store.
+     */
     class Compare {
         const DataStoreType &_store;
-        const EntryType _value;
+        const EntryType &_value;
 public:
         Compare(const DataStoreType &store, const EntryType &value)
             : _store(store),
               _value(value)
         {
         }
-        Compare(const DataStoreType &store)
-            : _store(store),
-              _value()
-        {
-        }
         inline const EntryType &get(EntryRef ref) const {
             if (ref.valid()) {
                 RefType iRef(ref);
-                return _store.template getBufferEntry<WrappedEntry>(iRef.bufferId(), iRef.offset())->value();
+                return *_store.template getBufferEntry<EntryType>(iRef.bufferId(), iRef.offset());
             } else {
                 return _value;
             }
@@ -61,24 +53,11 @@ public:
             return lhsValue < rhsValue;
         }
     };
-    class WrappedCompare {
-        const Compare &_comp;
-    public:
-        WrappedCompare(const Compare &comp)
-            : _comp(comp)
-        {
-        }
-        inline bool operator()(EntryRef lhs, EntryRef rhs) const
-        {
-            return _comp(lhs, rhs);
-        }
-    };
-
-    using UniqueStoreBufferType = BufferType<WrappedEntry>;
+    using UniqueStoreBufferType = BufferType<EntryType>;
     using DictionaryTraits = btree::BTreeTraits<32, 32, 7, true>;
     using Dictionary = btree::BTree<EntryRef, uint32_t,
                                     btree::NoAggregated,
-                                    const WrappedCompare,
+                                    Compare,
                                     DictionaryTraits>;
 private:
     DataStoreType _store;
@@ -87,17 +66,16 @@ private:
     Dictionary _dict;
     using generation_t = vespalib::GenerationHandler::generation_t;
 
-    const WrappedEntry &getWrapped(EntryRef ref) const
-    {
-        RefType iRef(ref);
-        return *_store.template getBufferEntry<WrappedEntry>(iRef.bufferId(), iRef.offset());
-    }
 public:
     UniqueStore();
     ~UniqueStore();
     EntryRef move(EntryRef ref);
-    EntryRef add(const EntryType &array);
-    const EntryType &get(EntryRef ref) const { return getWrapped(ref).value(); }
+    EntryRef add(const EntryType &value);
+    const EntryType &get(EntryRef ref) const
+    {
+        RefType iRef(ref);
+        return *_store.template getBufferEntry<EntryType>(iRef.bufferId(), iRef.offset());
+    }
     void remove(EntryRef ref);
     ICompactionContext::UP compactWorst();
     MemoryUsage getMemoryUsage() const;
@@ -107,6 +85,7 @@ public:
     void trimHoldLists(generation_t firstUsed) { _dict.getAllocator().trimHoldLists(firstUsed); _store.trimHoldLists(firstUsed); }
     vespalib::GenerationHolder &getGenerationHolder(void) { return _store.getGenerationHolder(); }
     void setInitializing(bool initializing) { _store.setInitializing(initializing); }
+    void freeze() { _dict.getAllocator().freeze(); }
 
     // Should only be used for unit testing
     const BufferState &bufferState(EntryRef ref) const;
